--- ui/user/src/routes/mcp-access-policies/+page.svelte
@@ -1,171 +1,0 @@
-<script lang="ts">
-	import { page } from '$app/state';
-	import Confirm from '$lib/components/Confirm.svelte';
-	import Layout from '$lib/components/Layout.svelte';
-	import AccessControlRuleForm from '$lib/components/admin/AccessControlRuleForm.svelte';
-	import IconButton from '$lib/components/primitives/IconButton.svelte';
-	import Table from '$lib/components/table/Table.svelte';
-	import { MCP_PUBLISHER_ALL_OPTION, PAGE_TRANSITION_DURATION } from '$lib/constants';
-	import { MCP_ACCESS_POLICY_FIELD_IDS } from '$lib/constants';
-	import {
-		fetchMcpServerAndEntries,
-		getPoweruserWorkspace,
-		initMcpServerAndEntries
-	} from '$lib/context/poweruserWorkspace.svelte';
-	import { UserService, type AccessControlRule } from '$lib/services';
-	import { clearUrlParams, goto } from '$lib/url';
-	import { openUrl } from '$lib/utils';
-	import { BookOpenText, Plus, Trash2 } from '@lucide/svelte';
-	import { onMount } from 'svelte';
-	import { fly } from 'svelte/transition';
-
-	let { data } = $props();
-	let { accessControlRules, workspaceId } = $derived(data);
-
-	initMcpServerAndEntries();
-
-	const mcpServersAndEntries = getPoweruserWorkspace();
-	let showCreateRule = $derived(page.url.searchParams.has('new'));
-	let ruleToDelete = $state<AccessControlRule>();
-
-	async function navigateToCreated(rule: AccessControlRule) {
-		clearUrlParams(['new']);
-		goto(`/mcp-access-policies/${rule.id}`);
-	}
-
-	const duration = PAGE_TRANSITION_DURATION;
-	const totalServers = $derived(
-		mcpServersAndEntries.entries.length + mcpServersAndEntries.servers.length
-	);
-
-	onMount(async () => {
-		if (workspaceId) {
-			fetchMcpServerAndEntries(workspaceId);
-		}
-	});
-
-	let title = $derived(showCreateRule ? 'Create MCP Access Policy' : 'MCP Access Policies');
-</script>
-
-<Layout {title} showBackButton={showCreateRule}>
-	{#snippet rightNavActions()}
-		{@render addRuleButton()}
-	{/snippet}
-	<div
-		class="h-full w-full"
-		in:fly={{ x: 100, duration, delay: duration }}
-		out:fly={{ x: -100, duration }}
-	>
-		{#if showCreateRule}
-			{@render createRuleScreen()}
-		{:else}
-			<div
-				class="flex flex-col gap-8"
-				in:fly={{ x: 100, delay: duration, duration }}
-				out:fly={{ x: -100, duration }}
-			>
-				{#if accessControlRules.length === 0}
-					<div class="mt-12 flex w-md flex-col items-center gap-4 self-center text-center">
-						<BookOpenText class="text-base-content/80 size-24 opacity-25" />
-						<h4 class="text-muted-content text-lg font-semibold">
-							No created MCP access policies.
-						</h4>
-						<p class="text-muted-content text-sm font-light">
-							Looks like you don't have any access policies created yet. <br />
-							Click the button below to get started.
-						</p>
-
-						{@render addRuleButton(MCP_ACCESS_POLICY_FIELD_IDS.addPolicyEmptyBtn)}
-					</div>
-				{:else}
-					<Table
-						data={accessControlRules}
-						fields={['displayName', 'servers']}
-						onClickRow={(d, isCtrlClick) => {
-							const url = `/mcp-access-policies/${d.id}`;
-							openUrl(url, isCtrlClick);
-						}}
-						headers={[
-							{
-								title: 'Name',
-								property: 'displayName'
-							}
-						]}
-					>
-						{#snippet actions(d)}
-							<IconButton
-								variant="danger"
-								onclick={(e) => {
-									e.stopPropagation();
-									ruleToDelete = d;
-								}}
-								tooltip={{ text: 'Delete Rule' }}
-							>
-								<Trash2 class="size-4" />
-							</IconButton>
-						{/snippet}
-						{#snippet onRenderColumn(property, d)}
-							{#if property === 'servers'}
-								{@const hasEverything = d.resources?.find((r) => r.id === '*')}
-								{@const count = hasEverything
-									? totalServers
-									: ((d.resources &&
-											d.resources.filter(
-												(r) => r.type === 'mcpServerCatalogEntry' || r.type === 'mcpServer'
-											).length) ??
-										0)}
-								{count ? count : '-'}
-							{:else}
-								{d[property as keyof typeof d]}
-							{/if}
-						{/snippet}
-					</Table>
-				{/if}
-			</div>
-		{/if}
-	</div>
-</Layout>
-
-{#snippet addRuleButton(id: string = MCP_ACCESS_POLICY_FIELD_IDS.addPolicyBtn)}
-	<button
-		{id}
-		class="btn btn-primary flex items-center gap-1 text-sm"
-		onclick={() => {
-			goto(`/mcp-access-policies?new=true`);
-		}}
-	>
-		<Plus class="size-4" /> Add Access Policy
-	</button>
-{/snippet}
-
-{#snippet createRuleScreen()}
-	<div
-		class="h-full w-full"
-		in:fly={{ x: 100, delay: duration, duration }}
-		out:fly={{ x: -100, duration }}
-	>
-		<AccessControlRuleForm
-			onCreate={navigateToCreated}
-			entity="workspace"
-			id={workspaceId}
-			mcpEntriesContextFn={getPoweruserWorkspace}
-			all={MCP_PUBLISHER_ALL_OPTION}
-		/>
-	</div>
-{/snippet}
-
-<Confirm
-	msg={`Delete ${ruleToDelete?.displayName || 'this access policy'}?`}
-	show={Boolean(ruleToDelete)}
-	onsuccess={async () => {
-		if (!ruleToDelete || !workspaceId) return;
-		await UserService.deleteWorkspaceAccessControlRule(workspaceId, ruleToDelete.id);
-		accessControlRules = await UserService.listWorkspaceAccessControlRules(workspaceId);
-		ruleToDelete = undefined;
-	}}
-	oncancel={() => (ruleToDelete = undefined)}
-/>
-
-<svelte:head>
-	<title>Obot | MCP Access Policies</title>
-</svelte:head>
