+++ ui/user/src/routes/mcp-access-policies/[id]/+page.ts
@@ -1,21 +1,6 @@
-import { handleRouteError } from '$lib/errors';
-import { UserService } from '$lib/services';
 import type { PageLoad } from './$types';
+import { redirect } from '@sveltejs/kit';
 
-export const load: PageLoad = async ({ params, fetch, parent }) => {
-	const { id } = params;
-	const { profile } = await parent();
-	let accessControlRule;
-	let workspaceId;
-	try {
-		workspaceId = await UserService.fetchWorkspaceIDForProfile(profile?.id, { fetch });
-		accessControlRule = await UserService.getWorkspaceAccessControlRule(workspaceId, id, { fetch });
-	} catch (err) {
-		handleRouteError(err, `/mcp-access-policies/${id}`, profile);
-	}
-
-	return {
-		accessControlRule,
-		workspaceId
-	};
+export const load: PageLoad = ({ params, url }) => {
+	throw redirect(301, `/mcp-servers/access-policies/${encodeURIComponent(params.id)}${url.search}`);
 };
