+++ ui/user/src/routes/mcp-catalog/c/[id]/+page.ts
@@ -1,22 +1,6 @@
-import { handleRouteError } from '$lib/errors';
-import { getMCPCatalogEntry } from '../../utils';
 import type { PageLoad } from './$types';
+import { redirect } from '@sveltejs/kit';
 
-export const load: PageLoad = async ({ params, url, fetch, parent }) => {
-	const { id } = params;
-	const wid = url.searchParams.get('wid');
-
-	const { profile } = await parent();
-	const prefix = profile.hasAdminAccess?.() ? '/admin' : '';
-
-	let catalogEntry;
-	try {
-		catalogEntry = await getMCPCatalogEntry(id, wid, profile, fetch);
-	} catch (err) {
-		handleRouteError(err, `${prefix}/mcp-catalog/c/${id}`, profile);
-	}
-
-	return {
-		catalogEntry
-	};
+export const load: PageLoad = ({ params, url }) => {
+	throw redirect(301, `/mcp-servers/c/${encodeURIComponent(params.id)}${url.search}`);
 };
