+++ ui/user/src/routes/mcp-catalog/c/[id]/instance/[ms_id]/+page.ts
@@ -1,30 +1,9 @@
-import { handleRouteError } from '$lib/errors';
-import { getMCPCatalogEntry, getSingleOrRemoteMcpServer } from '../../../../utils';
 import type { PageLoad } from './$types';
+import { redirect } from '@sveltejs/kit';
 
-export const load: PageLoad = async ({ params, url, fetch, parent }) => {
-	const catalogEntryId = params.id;
-	const mcpServerId = params.ms_id;
-	const { profile } = await parent();
-	const prefix = profile.hasAdminAccess?.() ? '/admin' : '';
-	const wid = url.searchParams.get('wid');
-
-	let catalogEntry;
-	let mcpServer;
-	try {
-		catalogEntry = await getMCPCatalogEntry(catalogEntryId, wid, profile, fetch);
-		mcpServer = await getSingleOrRemoteMcpServer(mcpServerId, catalogEntryId, wid, profile, fetch);
-	} catch (err) {
-		handleRouteError(
-			err,
-			`${prefix}/mcp-catalog/c/${catalogEntryId}/instance/${mcpServerId}`,
-			profile
-		);
-	}
-
-	return {
-		catalogEntry,
-		mcpServerId,
-		mcpServer
-	};
+export const load: PageLoad = ({ params, url }) => {
+	throw redirect(
+		301,
+		`/mcp-servers/c/${encodeURIComponent(params.id)}/instance/${encodeURIComponent(params.ms_id)}${url.search}`
+	);
 };
