+++ ui/user/src/routes/mcp-catalog/s/[id]/+page.ts
@@ -1,36 +1,6 @@
-import { handleRouteError, parseErrorContent } from '$lib/errors';
-import { getMCPCatalogEntry, getMCPCatalogServer } from '../../utils';
 import type { PageLoad } from './$types';
+import { redirect } from '@sveltejs/kit';
 
-export const load: PageLoad = async ({ params, url, fetch, parent }) => {
-	const { profile } = await parent();
-	const { id } = params;
-	const wid = url.searchParams.get('wid');
-
-	const prefix = profile.hasAdminAccess?.() ? '/admin' : '';
-
-	let mcpServer;
-	try {
-		mcpServer = await getMCPCatalogServer(id, wid, profile, fetch);
-	} catch (err) {
-		handleRouteError(err, `${prefix}/mcp-catalog/s/${id}`, profile);
-	}
-
-	let catalogEntry;
-	if (mcpServer?.catalogEntryID) {
-		try {
-			catalogEntry = await getMCPCatalogEntry(mcpServer.catalogEntryID, wid, profile, fetch);
-		} catch (err) {
-			// Only swallow 404 — the referenced entry was deleted but the server still
-			// points at it. Surface anything else so real failures aren't hidden.
-			if (parseErrorContent(err).status !== 404) {
-				handleRouteError(err, `${prefix}/mcp-catalog/s/${id}`, profile);
-			}
-		}
-	}
-
-	return {
-		mcpServer,
-		catalogEntry
-	};
+export const load: PageLoad = ({ params, url }) => {
+	throw redirect(301, `/mcp-servers/s/${encodeURIComponent(params.id)}${url.search}`);
 };
