+++ ui/user/src/routes/mcp-servers/access-policies/[id]/+page.ts
@@ -1,19 +1,29 @@
 import { handleRouteError } from '$lib/errors';
-import { AdminService } from '$lib/services';
-import { profile } from '$lib/stores';
+import { AdminService, UserService } from '$lib/services';
 import type { PageLoad } from './$types';
 
-export const load: PageLoad = async ({ params, fetch }) => {
+export const load: PageLoad = async ({ params, fetch, parent }) => {
 	const { id } = params;
+	const { profile } = await parent();
 
 	let accessControlRule;
+	let workspaceId: string | undefined;
+
 	try {
-		accessControlRule = await AdminService.getAccessControlRule(id, { fetch });
+		if (profile.hasAdminAccess?.()) {
+			accessControlRule = await AdminService.getAccessControlRule(id, { fetch });
+		} else {
+			workspaceId = await UserService.fetchWorkspaceIDForProfile(profile?.id, { fetch });
+			accessControlRule = await UserService.getWorkspaceAccessControlRule(workspaceId, id, {
+				fetch
+			});
+		}
 	} catch (err) {
-		handleRouteError(err, `/admin/mcp-access-policies/${id}`, profile.current);
+		handleRouteError(err, `/mcp-servers/access-policies/${id}`, profile);
 	}
 
 	return {
-		accessControlRule
+		accessControlRule,
+		workspaceId
 	};
 };
