+++ ui/user/src/routes/mcp-servers/access-policies/w/[wid]/r/[id]/+page.svelte
@@ -18,7 +18,7 @@
 		<AccessControlRuleForm
 			{accessControlRule}
 			onUpdate={() => {
-				goto('/admin/mcp-access-policies');
+				goto('/mcp-servers?view=access-policy');
 			}}
 			entity="workspace"
 			id={workspaceId}
