+++ ui/user/src/routes/mcp-servers/c/[id]/+page.svelte
