+++ ui/user/src/routes/mcp-servers/c/[id]/instance/[ms_id]/+page.svelte
