+++ ui/user/src/routes/mcp-servers/c/[id]/instance/[ms_id]/details/page.svelte.spec.ts
