+++ ui/user/src/routes/mcp-servers/s/[id]/+page.svelte
