+++ ui/user/src/routes/mcp-servers/s/[id]/details/page.svelte.spec.ts
